add login

+++ app/Http/Controllers/TasksController.php
@@ -17,14 +17,38 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-         $tasks = Task::all();
+    {    
+        
+        //$tasks = Task::all();
 
         // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        //return view('tasks.index', [
+        //'tasks' => $tasks,
+        //]);
+    
+    
+    
+       
+       $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+                
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data, );
+        
+   
     }
+       
         //
 
 
@@ -41,6 +65,7 @@ class TasksController extends Controller
         // メッセージ作成ビューを表示
         return view('tasks.create', [
             'task' => $task,
+            
         ]);
          //
     }
@@ -59,16 +84,22 @@ class TasksController extends Controller
             'status' => 'required|max:10',
               
         ]);
-       // メッセージを作成
+       
+       
+        
+        
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
+        $task->user_id = \Auth::id();
         $task->save();
+       
         
-
+       //  return back();
+        
         // トップページへリダイレクトさせる
         return redirect('/');
-        //
+        
     }
 
     /**
@@ -78,7 +109,7 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show($id)
-    {
+     {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
@@ -88,7 +119,6 @@ class TasksController extends Controller
         ]);
         //
     }
-
     /**
      * Show the form for editing the specified resource.
      *
@@ -96,7 +126,7 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {
+   {
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
@@ -106,6 +136,8 @@ class TasksController extends Controller
         ]);
         //
     }
+
+
 
     /**
      * Update the specified resource in storage.
@@ -128,7 +160,7 @@ class TasksController extends Controller
         $task->status = $request->status;    
         $task->content = $request->content;
         $task->save();
-
+        
         // トップページへリダイレクトさせる
         return redirect('/');
         //
@@ -142,14 +174,16 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        //
-        
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
+        
+         return redirect('/');
     }
 }
